refactor skipstate
