+++ src/levanter/optim/config.py
@@ -3,7 +3,7 @@ import re
 import warnings
 from dataclasses import dataclass
 from functools import partial
-from typing import NamedTuple, Optional
+from typing import Optional
 
 import draccus
 import equinox as eqx
@@ -15,15 +15,8 @@ from jax import numpy as jnp
 import haliax
 
 import levanter.tracker
+from levanter.optim.skipstep import SkipStepConfig
 from levanter.utils.jax_utils import leaf_key_paths
-
-
-class SkipStepState(NamedTuple):
-    inner_opt_state: optax.OptState
-    losses: jax.Array  # Shape: (rolling_interval_length + 1,), dtype: jnp.float32
-    grad_norms: jax.Array  # Shape: (rolling_interval_length + 1,), dtype: jnp.float32
-    current_idx: jax.Array  # Scalar int
-    count: jax.Array  # Scalar int
 
 
 @dataclass
@@ -55,9 +48,6 @@ class OptimizerConfig(draccus.ChoiceRegistry, abc.ABC):
     default_weight_decay_mask: Optional[bool] = None
     """Whether to apply a default reasonable weight decay to modules not explicitly masked. None means it will if
     no weight_decay_modules are set. False means it will not. True means it will regardless of weight_decay_modules."""
-
-    skip_step: Optional["SkipStepConfig"] = None
-    """If set, defines the configuration for skipping steps when gradients are too large."""
 
     @classmethod
     def default_choice_name(cls) -> Optional[str]:
@@ -304,122 +294,6 @@ class HessianOptConfig(OptimizerConfig, abc.ABC):
     """How often to update the hessian approximation."""
 
 
-@dataclass
-class SkipStepConfig:
-    rolling_interval_length: int = 128
-    sigma_factor: float = 6.0
-
-    def wrap(self, inner_optimizer: optax.GradientTransformation) -> optax.GradientTransformation:
-        def init_fn(params):
-            inner_opt_state = inner_optimizer.init(params)
-            losses = jnp.full((self.rolling_interval_length + 1,), jnp.nan, dtype=jnp.float32)
-            grad_norms = jnp.full((self.rolling_interval_length + 1,), jnp.nan, dtype=jnp.float32)
-            current_idx = jnp.zeros((), dtype=jnp.int32)
-            count = jnp.zeros((), dtype=jnp.int32)
-            return SkipStepState(
-                inner_opt_state=inner_opt_state,
-                losses=losses,
-                grad_norms=grad_norms,
-                current_idx=current_idx,
-                count=count,
-            )
-
-        def update_fn(updates, state: SkipStepState, params, *, loss: Optional[jax.Array] = None, **extra_args):
-            if loss is None:
-                raise ValueError(
-                    "Loss must be provided as a keyword argument to the update function when using SkipStepOptimizer."
-                )
-
-            global_norm = optax.global_norm(updates)
-
-            # Update rolling statistics (circular buffer logic)
-            new_losses = state.losses.at[state.current_idx].set(loss)
-            new_grad_norms = state.grad_norms.at[state.current_idx].set(global_norm)
-            new_current_idx = (state.current_idx + 1) % (self.rolling_interval_length + 1)
-            new_count = jnp.minimum(state.count + 1, self.rolling_interval_length + 1)
-
-            # Decision Logic (step_factor)
-            min_data_points = jnp.maximum(2, self.rolling_interval_length // 2)
-            can_skip = new_count >= min_data_points
-
-            # Create a mask for valid (non-NaN) entries
-            idx_range = jnp.arange(self.rolling_interval_length + 1)
-            valid_mask = idx_range < new_count
-
-            # Calculate stats for skipping, ensuring NaNs are handled and std is not zero
-            masked_losses = jnp.where(valid_mask, new_losses, jnp.nan)
-            loss_mean = jnp.nanmean(masked_losses)
-            loss_std = jnp.nanstd(masked_losses)
-            loss_std_safe = jnp.maximum(loss_std, 1e-6)
-
-            masked_grad_norms = jnp.where(valid_mask, new_grad_norms, jnp.nan)
-            grad_norm_mean = jnp.nanmean(masked_grad_norms)
-            grad_norm_std = jnp.nanstd(masked_grad_norms)
-            grad_norm_std_safe = jnp.maximum(grad_norm_std, 1e-6)
-            
-            # if mean or std is nan (e.g. if new_count is 0 or 1), then don't skip
-            # this is a bit of a hack, but it's fine because can_skip will be false anyway in early steps
-            loss_mean = jnp.nan_to_num(loss_mean, nan=0.0)
-            loss_std_safe = jnp.nan_to_num(loss_std_safe, nan=1e-6)
-            grad_norm_mean = jnp.nan_to_num(grad_norm_mean, nan=0.0)
-            grad_norm_std_safe = jnp.nan_to_num(grad_norm_std_safe, nan=1e-6)
-
-
-            loss_threshold = loss_mean + self.sigma_factor * loss_std_safe
-            grad_norm_threshold = grad_norm_mean + self.sigma_factor * grad_norm_std_safe
-
-            loss_ok = loss <= loss_threshold
-            grad_norm_ok = global_norm <= grad_norm_threshold
-            
-            # we skip if loss OR grad_norm are NOT ok.
-            # so should_skip is true if NOT (loss_ok AND grad_norm_ok)
-            # which is (NOT loss_ok) OR (NOT grad_norm_ok)
-            should_skip = jnp.logical_not(jnp.logical_and(loss_ok, grad_norm_ok))
-
-            step_factor = jnp.where(jnp.logical_and(can_skip, should_skip), 0.0, 1.0)
-
-            # Apply Update based on step_factor
-            # We need to define the two branches for jax.lax.cond
-            # Signature for cond branches: operand -> result
-            # Here, operand is (updates, state.inner_opt_state, params, extra_args)
-            # Result is (new_inner_updates, new_inner_opt_state)
-
-            def skip_step_branch(_):
-                # updates are zeroed out, inner state is preserved
-                zeroed_updates = jax.tree_util.tree_map(jnp.zeros_like, updates)
-                return zeroed_updates, state.inner_opt_state
-
-            def proceed_step_branch(_):
-                # updates are applied as is, inner state is updated
-                return inner_optimizer.update(updates, state.inner_opt_state, params, **extra_args)
-
-            # Use jax.lax.cond to choose branch based on step_factor
-            # pred is true if step_factor == 0.0 (skip)
-            new_inner_updates, new_inner_opt_state = jax.lax.cond(
-                step_factor == 0.0,
-                skip_step_branch,
-                proceed_step_branch,
-                operand=None  # operand not strictly needed as branches capture necessary values
-            )
-
-            # Log whether a step was skipped
-            # TODO: make this a proper metric when levanter.tracker.log_metrics is a thing
-            # For now, we can log it as a hyperparameter, though it's not ideal.
-            # Alternatively, print it, but that's noisy.
-            # For now, let's skip explicit logging here as it's not requested and might be noisy.
-            # A more advanced solution would involve passing a logging callback or similar.
-
-            return new_inner_updates, SkipStepState(
-                inner_opt_state=new_inner_opt_state,
-                losses=new_losses,
-                grad_norms=new_grad_norms,
-                current_idx=new_current_idx,
-                count=new_count,
-            )
-
-        return optax.GradientTransformation(init_fn, update_fn)
-
-
 @OptimizerConfig.register_subclass("adam")
 @dataclass
 class AdamConfig(OptimizerConfig):
@@ -429,6 +303,9 @@ class AdamConfig(OptimizerConfig):
     beta2: float = 0.95
     epsilon: float = 1e-8
     max_grad_norm: Optional[float] = 1.0
+
+    skip_step: Optional[SkipStepConfig] = None
+    """If set, defines the configuration for skipping steps when gradients are too large."""
 
     def build(self, num_train_steps):
         """Creates the optimizer"""
